backend/requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
gunicorn==21.2.0
mysql-connector-python==8.0.33
flask-cors